Create README.md
This is just a pointless car you can build with an arduino with two dc motors, a servo with an ultrasonic sensor attached to it. It moves forward until it is within 5 cm from an object if something is within 5 cm then the servo with the ultrasonic sensor on it with it will turn to the left. If there is nothing on the left side then the car will turn to the left and move forward. If there is something on the left side then the servo will turn right. If there is nothing on the right side then the car will turn right and move forward. If there is something on the right side then the car will turn 180 degrees and return the way it came.