Add files via upload
"Finished Product" was just a project I did in high school that began a much more interesting project that took me a good while to finish. It is a very simple "lock box" which really was just pressing buttons to lock or unlock a servo lock. It uses push buttons, a remote, a number pad, and a keypad. It also has an alarm system that I had intended to be if you entered an incorrect password too many times it would go off and a sound sensor that if the alarm went off it would disconnect the servo. However I couldn't figure out how to make passwords so it turned in to if you press the wrong button the alarm went off. In this version I couldn't figure out how to make any combinations for passwords but after working on it for a while and doing quite a bit of research I figured it out and am still just adding random stuff to this lock box project that this project started.