Add files via upload
This is the last working version of this project but I am still adding points of entry and attempting to simplify the array checking system. It is a project I have been working on for a while to build a lockbox controlled by an arduino it currently has three points of entry but I am currently adding a rotary encoder to work like a combination lock, an rfid card, and an LCD screen to show the position of the rotary encoder, incorrect passwords etc.